requirements.txt
absl-py==0.9.0
altgraph==0.17
attrs==19.3.0
audiolazy==0.6
audioread==2.1.8
backcall==0.1.0
beautifulsoup4==4.9.1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1588608214987/work
brotlipy==0.7.0
build==0.0.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
cupy-cuda101==7.2.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
fastrlock==0.4
future==0.18.2
grpcio==1.23.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
ipykernel @ file:///D:/bld/ipykernel_1590023017759/work/dist/ipykernel-5.3.0-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1590797082149/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
Js2Py==0.70
json5==0.9.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1590229434073/work
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.32.1
macholib==1.11
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1589366472132/work
MarkupSafe==1.1.1
matlabengineforpython===R2018a
matplotlib==3.2.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.6
notebook @ file:///D:/bld/notebook_1588887315004/work
numba==0.49.1
numpy==1.18.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pefile==2019.4.18
pep517==0.8.2
pickleshare==0.7.5
Pillow==7.0.0
pipwin==0.5.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1590412252446/work
prompt-toolkit==3.0.5
protobuf==3.12.1
PyAudio @ file:///C:/Users/WilliamLambert/pipwin/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycparser==2.20
pycryptodome==3.8.2
Pygments==2.6.1
PyInstaller==3.6
pyjsparser==2.7.1
pynvml==8.0.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPrind==2.11.2
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.16.0
pySmartDL==1.3.3
PySocks==1.7.1
pystoi==0.3.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole @ file:///home/conda/feedstock_root/build_artifacts/qtconsole_1589392299544/work
QtPy==1.9.0
requests==2.23.0
resampy==0.2.2
scikit-learn==0.23.0
scipy==1.4.1
Send2Trash==1.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sklearn==0.0
sounddevice==0.3.15
SoundFile==0.10.3.post1
soupsieve==2.0.1
tensorboard==2.0.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
toml==0.10.1
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.1.0
